requirements.txt
streamlit
langgraph
langchain
langchain-groq
python-docx
pdfplumber
openai
python-dotenv
